rename invitations service
